Tabs are sorted by type
The mode of operation has not changed. However, with the new update, the first two characters are missing in the display of the "Device Type". This can be corrected by editing the device and selecting the same type (router, switch, ...) again. The background is that I insert numbers at the beginning to be able in the representation of the tabs the devices by type. Previously, the devices were "sorted" according to the order of creation. Now by Router, Switch, WLAN and Powerline