Addressing the issue https://github.com/leiweibau/Pi.Alert/issues/67
While it was possible to assign a port to a switch on a router, which was displayed in the router tab, it was not possible to assign multiple switches to a router, so that the router was displayed on all switch tabs. Now it is possible to manually configure ports when configuring a switch on the network page. Documentation on Help/FAQ will follow