WebGUI Notification: Badge in the upper right corner
Displays the notifications now also on the web page. The indication is a small red badge above the Pi.Alert icon, in the upper right corner. Clicking on the icon opens a menu with a button that leads to a page with all notifications. These can now be viewed and deleted individually. The deletion of all notifications will be implemented later. The badge is displayed as long as notifications are present.
This can be disabled in the pialert.conf (REPORT_WEBGUI = False)