Pi.Alert GUI Journal 01
Configuration changes or work on the database via the web interface are recorded in an application journal to provide some form of security and traceability.

The journal works with codes that become readable through the language files, so the journal is also multilingual.

The journal is not yet included in the installation package.

The work is still at an early stage and will be continued.

https://github.com/leiweibau/Pi.Alert/issues/140