Changed pialert-cli
The functions "set_permissions" and "set_sudoers" have been changed.

set_permissions:
There are now 3 optional switches
--lxc: sets the username to "root
--custom: an individual user name can be entered (only for experienced users)
--homedir: the username is determined from the parent home directory
If none of the switches is specified, only the group permissions will be adjusted.

set_sudoers:
There are now 2 optional switches
--lxc: sets the username to "root
--custom: an individual username can be entered (only for experienced users)
If none of the switches is specified, the username will be determined from the parent home directory