Install Ookla speedtest client
Implemented a method for architecture detection, downloading and unpacking the archive.

At the moment the archive is downloaded only if the system platform matches one of the supported platforms. Compatible platforms are not yet covered (amd64 can run i386 or x86_64).

Regarding to https://github.com/leiweibau/Pi.Alert/issues/160