Journal improvements
In some cases, the journal cannot write to the database because it is being used by another process. To ensure that the journal entries are not lost, a buffer has now been introduced which is used if an entry cannot be written. If the next journal entry is to be written, the buffer is extended in the event of an error or, if successful, the buffer is processed so that older entries appear in the journal afterwards.