Add he option to use only the beginning of a MAC address
If several Mac addresses with the same beginning are to be filtered out, it is no longer necessary to enter all the individual addresses.

The addresses:
aa:bb:cc:12:34:56
aa:bb:c0:22:24:26
aa:bb:ce:a2:a3:a4

can now be filtered out with the entry 'aa:bb:c'.

However, this only works with the beginning of the MAC address. Addresses that end identically cannot be filtered out in this way.

Regarding to https://github.com/leiweibau/Pi.Alert/issues/246