Auto Update check 01
Thanks to @jokob-sk for the inspiration.

Initial work on the implementation of an automatic update check twice a day. A configuration of the number and timing is not yet planned at this point. The notification of an update is initially only planned in the frontend. Notifications via email or the configurable push services may follow later.